Start to extract d3 functionality into self-contained file
